BIG CHANGES
sorted functions into files
