+++ SeniorProject2025/SeniorProject2025.vcxproj.filters
@@ -24,19 +24,31 @@
     <ClCompile Include="flags.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="directivesAndDataErrors.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="calculations.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="arm_operators.h">
-      <Filter>Source Files</Filter>
+    <ClInclude Include="main.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="calculations.h">
-      <Filter>Source Files</Filter>
+    <ClInclude Include="directivesAndDataErrors.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="arm_operators.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="branchAndSubroutines.h">
-      <Filter>Source Files</Filter>
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="calculations.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="flags.h">
-      <Filter>Source Files</Filter>
+      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
